+++ analytics/graph_record_test.go
@@ -1,0 +1,227 @@
+package analytics
+
+import (
+	"encoding/base64"
+	"fmt"
+	"github.com/google/go-cmp/cmp"
+	"github.com/google/go-cmp/cmp/cmpopts"
+	"github.com/stretchr/testify/assert"
+	"testing"
+	"time"
+)
+
+func TestAnalyticsRecord_ToGraphRecord(t *testing.T) {
+	recordSample := AnalyticsRecord{
+		TimeStamp:    time.Date(2022, 1, 1, 0, 0, 0, 0, time.UTC),
+		Method:       "POST",
+		Host:         "localhost:8281",
+		Path:         "/",
+		RawPath:      "/",
+		APIName:      "test-api",
+		APIID:        "test-api",
+		ApiSchema:    base64.StdEncoding.EncodeToString([]byte(sampleSchema)),
+		ResponseCode: 200,
+		Day:          1,
+		Month:        1,
+		Year:         2022,
+		Hour:         0,
+	}
+	graphRecordSample := GraphRecord{
+		AnalyticsRecord: recordSample,
+		Types:           make(map[string][]string),
+	}
+
+	testCases := []struct {
+		expected     func(string, string) GraphRecord
+		modifyRecord func(a AnalyticsRecord) AnalyticsRecord
+		title        string
+		request      string
+		response     string
+		expectedErr  string
+	}{
+		{
+			title:    "no error",
+			request:  `{"query":"query{\n  characters(filter: {\n    \n  }){\n    info{\n      count\n    }\n  }\n}"}`,
+			response: `{"data":{"characters":{"info":{"count":758}}}}`,
+			expected: func(request, response string) GraphRecord {
+				g := graphRecordSample
+				g.HasErrors = false
+				g.Types = map[string][]string{
+					"Characters": {"info"},
+					"Info":       {"count"},
+				}
+				g.OperationType = "Query"
+				return g
+			},
+		},
+		{
+			title:    "no error mutation",
+			request:  `{"query":"mutation{\n  changeCharacter()\n}"}`,
+			response: `{"data":{"characters":{"info":{"count":758}}}}`,
+			expected: func(request, response string) GraphRecord {
+				g := graphRecordSample
+				g.HasErrors = false
+				g.OperationType = "Mutation"
+				return g
+			},
+		},
+		{
+			title:    "no error subscription",
+			request:  `{"query":"subscription{\n  listenCharacter(){\n    info{\n      count\n    }\n  }\n}"}`,
+			response: `{"data":{"characters":{"info":{"count":758}}}}`,
+			expected: func(request, response string) GraphRecord {
+				g := graphRecordSample
+				g.HasErrors = false
+				g.Types = map[string][]string{
+					"Characters": {"info"},
+					"Info":       {"count"},
+				}
+				g.OperationType = "Subscription"
+				return g
+			},
+		},
+		{
+			title:    "bad document",
+			request:  `{"query":"subscriptiona{\n  listenCharacter(){\n    info{\n      count\n    }\n  }\n}"}`,
+			response: `{"data":{"characters":{"info":{"count":758}}}}`,
+			expected: func(request, response string) GraphRecord {
+				return GraphRecord{}
+			},
+			expectedErr: "error generating documents",
+		},
+		{
+			title:    "no error list operation",
+			request:  `{"query":"query{\n  listCharacters(){\n    info{\n      count\n    }\n  }\n}"}`,
+			response: `{"data":{"characters":{"info":{"count":758}}}}`,
+			expected: func(request, response string) GraphRecord {
+				g := graphRecordSample
+				g.HasErrors = false
+				g.Types = map[string][]string{
+					"Characters": {"info"},
+					"Info":       {"count"},
+				}
+				g.OperationType = "Query"
+				return g
+			},
+		},
+		{
+			title:    "has variables",
+			request:  `{"query":"query{\n  characters(filter: {\n    \n  }){\n    info{\n      count\n    }\n  }\n}","variables":{"a":"test"}}`,
+			response: `{"data":{"characters":{"info":{"count":758}}}}`,
+			expected: func(request, response string) GraphRecord {
+				g := graphRecordSample
+				g.HasErrors = false
+				g.Types = map[string][]string{
+					"Characters": {"info"},
+					"Info":       {"count"},
+				}
+				g.OperationType = "Query"
+				g.Variables = base64.StdEncoding.EncodeToString([]byte(`{"a":"test"}`))
+				return g
+			},
+		},
+		{
+			title:    "no operation",
+			request:  `{"query":"query main {\ncharacters {\ninfo\n}\n}\n\nquery second {\nlistCharacters{\ninfo\n}\n}","variables":null,"operationName":""}`,
+			response: `{"data":{"characters":{"info":{"count":758}}}}`,
+			expected: func(request, response string) GraphRecord {
+				return GraphRecord{}
+			},
+			expectedErr: "no operation name specified",
+		},
+		{
+			title:    "operation name specified",
+			request:  `{"query":"query main {\ncharacters {\ninfo\n}\n}\n\nquery second {\nlistCharacters{\ninfo\n secondInfo}\n}","variables":null,"operationName":"second"}`,
+			response: `{"data":{"characters":{"info":{"count":758}}}}`,
+			expected: func(request, response string) GraphRecord {
+				g := graphRecordSample
+				g.HasErrors = false
+				g.Types = map[string][]string{
+					"Characters": {"info", "secondInfo"},
+				}
+				g.OperationType = "Query"
+				return g
+			},
+			expectedErr: "",
+		},
+		{
+			title:   "has errors",
+			request: `{"query":"query{\n  characters(filter: {\n    \n  }){\n    info{\n      count\n    }\n  }\n}"}`,
+			response: `{
+  "errors": [
+    {
+      "message": "Name for character with ID 1002 could not be fetched.",
+      "locations": [{ "line": 6, "column": 7 }],
+      "path": ["hero", "heroFriends", 1, "name"]
+    }
+  ]
+}`,
+			expected: func(request, response string) GraphRecord {
+				g := graphRecordSample
+				g.HasErrors = true
+				g.Types = map[string][]string{
+					"Characters": {"info"},
+					"Info":       {"count"},
+				}
+				g.OperationType = "Query"
+				g.Errors = append(g.Errors, GraphError{
+					Message: "Name for character with ID 1002 could not be fetched.",
+					Path:    []interface{}{"hero", "heroFriends", float64(1), "name"},
+				})
+				return g
+			},
+		},
+		{
+			title: "corrupted raw request should error out",
+			modifyRecord: func(a AnalyticsRecord) AnalyticsRecord {
+				a.RawRequest = "this isn't a base64 is it?"
+				return a
+			},
+			expectedErr: "error decoding raw request",
+			expected: func(s, s2 string) GraphRecord {
+				return GraphRecord{}
+			},
+		},
+		{
+			title: "corrupted schema should error out",
+			modifyRecord: func(a AnalyticsRecord) AnalyticsRecord {
+				a.ApiSchema = "this isn't a base64 is it?"
+				return a
+			},
+			expectedErr: "error decoding schema",
+			expected: func(s, s2 string) GraphRecord {
+				return GraphRecord{}
+			},
+		},
+	}
+
+	for _, testCase := range testCases {
+		t.Run(testCase.title, func(t *testing.T) {
+			a := recordSample
+			a.RawRequest = base64.StdEncoding.EncodeToString([]byte(fmt.Sprintf(
+				requestTemplate,
+				len(testCase.request),
+				testCase.request,
+			)))
+			a.RawResponse = base64.StdEncoding.EncodeToString([]byte(fmt.Sprintf(
+				responseTemplate,
+				len(testCase.response),
+				testCase.response,
+			)))
+			if testCase.modifyRecord != nil {
+				a = testCase.modifyRecord(a)
+			}
+			expected := testCase.expected(testCase.request, testCase.response)
+			expected.AnalyticsRecord = a
+			gotten, err := a.ToGraphRecord()
+			if testCase.expectedErr != "" {
+				assert.ErrorContains(t, err, testCase.expectedErr)
+				return
+			}
+			assert.NoError(t, err)
+			if diff := cmp.Diff(expected, gotten, cmpopts.IgnoreFields(GraphRecord{}, "RawRequest", "RawResponse")); diff != "" {
+				t.Fatal(diff)
+			}
+		})
+	}
+}
